# docs/architecture/openapi.yaml
openapi: 3.1.0
info:
  title: Content Intelligence Platform API
  version: "0.1.0-mvp"
paths:
  /api/profiles:
    get:
      summary: Lista perfiles editoriales disponibles
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/ProfileDto" }
    post:
      summary: Registra un nuevo perfil (el .md ya debe existir en knowledge/profiles)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProfileRequest" }
      responses:
        "201": { description: Creado }
        "400": { $ref: "#/components/responses/ProblemDetails" }

  /api/generations:
    get:
      summary: Historial de generaciones
      parameters:
        - name: profileSlug
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/GenerationSummaryDto" }
    post:
      summary: Ejecuta el pipeline completo (ingesta + IA) y produce un guion
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/GenerateScriptRequest" }
      responses:
        "201":
          description: Guion generado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerationDetailDto" }
        "400": { $ref: "#/components/responses/ProblemDetails" }
        "422": { $ref: "#/components/responses/ProblemDetails" }

  /api/generations/{id}:
    get:
      summary: Detalle de una generación
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/GenerationDetailDto" }
        "404": { $ref: "#/components/responses/ProblemDetails" }

  /api/generations/{id}/rating:
    patch:
      summary: Califica una generación (para la mejora continua de sección 15)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RateGenerationRequest" }
      responses:
        "204": { description: Actualizado }
        "404": { $ref: "#/components/responses/ProblemDetails" }

components:
  responses:
    ProblemDetails:
      description: Error (RFC 7807)
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
    ProblemDetails:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
    ProfileDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        parentSlug: { type: string, nullable: true }
        version: { type: integer }
    CreateProfileRequest:
      type: object
      required: [slug, name, filePath, version]
      properties:
        slug: { type: string }
        name: { type: string }
        parentSlug: { type: string, nullable: true }
        filePath: { type: string }
        version: { type: integer }
    GenerateScriptRequest:
      type: object
      required: [sourceText, profileSlug, outputFormat]
      properties:
        sourceText: { type: string }
        profileSlug: { type: string }
        outputFormat:
          type: string
          enum: [TikTok, YouTubeShorts, InstagramReel, Podcast]
    GenerationSummaryDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        profileSlug: { type: string }
        createdAt: { type: string, format: date-time }
        rating: { type: integer, nullable: true }
    GenerationDetailDto:
      type: object
      properties:
        id: { type: string, format: uuid }
        profileSlug: { type: string }
        providerName: { type: string }
        resultMarkdown: { type: string }
        resultJson: { type: string }
        confidence: { type: number }
        missingInformation: { type: string, nullable: true }
        tokensInput: { type: integer }
        tokensOutput: { type: integer }
        costUsd: { type: number }
        createdAt: { type: string, format: date-time }
    RateGenerationRequest:
      type: object
      required: [rating]
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comments: { type: string, nullable: true }